move some components
