+++ src/ui/wordsCountBlock/WordsCountBlock.tsx
