+++ src/ui/wordsCountBlock/index.ts
